Merging dev into master

+++ app/analyzers/indicators/ma_ribbon.py
@@ -13,7 +13,8 @@ class MARibbon(IndicatorUtils):
 
     # Exponential Moving Average
     def EMA(self, df, n, field='close'):
-        return pandas.Series(talib.EMA(df[field].astype('f8').values, n), name='EMA_' + field.upper() + '_' + str(n), index=df.index)
+        return pandas.Series(talib.EMA(df[field].astype(
+            'f8').values, n), name='EMA_' + field.upper() + '_' + str(n), index=df.index)
 
     def MA_RIBBON(self, df, ma_series):
         ma_array = np.zeros([len(df), len(ma_series)])
@@ -36,13 +37,14 @@ class MARibbon(IndicatorUtils):
                     ma_array[idy, :], range(len(ma_series), 0, -1))
                 dist[idy] = max(ma_array[idy, :]) - min(ma_array[idy, :])
 
-        corr_ts = pandas.Series(corr*100, index=df.index,
+        corr_ts = pandas.Series(corr * 100, index=df.index,
                                 name="MARIBBON_CORR").round(2)
-        pval_ts = pandas.Series(pval*100, index=df.index,
+        pval_ts = pandas.Series(pval * 100, index=df.index,
                                 name="MARIBBON_PVAL").round(2)
         dist_ts = pandas.Series(dist, index=df.index, name="MARIBBON_DIST")
 
-        return pandas.concat([corr_ts, pval_ts, dist_ts] + ema_list, join='outer', axis=1)
+        return pandas.concat([corr_ts, pval_ts, dist_ts] +
+                             ema_list, join='outer', axis=1)
 
     def analyze(self, historical_data, pval_th, ma_series, signal=['ma_ribbon'],
                 hot_thresh=10, cold_thresh=-10):
@@ -54,7 +56,7 @@ class MARibbon(IndicatorUtils):
             pval_th (integer):
             hot_thresh (integer):
             cold_thresh (integer):
-            ma_series (list): 
+            ma_series (list):
 
         Returns:
             pandas.DataFrame: A dataframe containing the indicator and hot/cold values.
